refactor(mcp): address remaining Greptile review on exec tests
- Export createExecInnerToolCallResolver from tools/exec.ts. mcp.ts and
  the e2e emission test now both import the same factory instead of
  carrying parallel copies of the resolver lambda. The test now exercises
  the real production resolver; future guards added to the resolver
  automatically flow into the test.
- Parameterise the parseExecCallInnerToolName tests via it.each with a
  single 13-row truth table. Each input becomes a discrete test case so
  failures are easy to localize (instead of "this it block failed
  somewhere").

Generated-By: PostHog Code
Task-Id: 09ea33de-4e1a-4e91-b838-a05c834f4ced
